+++ packages/terra-form-select/src/combobox/Frame.jsx
@@ -199,7 +199,6 @@ class Frame extends React.Component {
     this.visuallyHiddenComponent = React.createRef();
     this.setSelectMenuRef = this.setSelectMenuRef.bind(this);
     this.shouldFocusDropdown = false;
-    this.hasEscPressed = false;
   }
 
   componentDidMount() {
@@ -272,6 +271,7 @@ class Frame extends React.Component {
   handleKeyDown(event) {
     const { intl } = this.props;
     const { keyCode, target } = event;
+    const hasValue = !!((this.state.searchValue.trim().length || this.props.value));
 
     if (keyCode === KeyCode.KEY_SPACE && target !== this.input) {
       event.preventDefault();
@@ -280,21 +280,9 @@ class Frame extends React.Component {
       event.preventDefault();
       this.openDropdown(event);
     } else if (this.state.isOpen && keyCode === KeyCode.KEY_ESCAPE) {
-      this.hasEscPressed = true;
       event.stopPropagation();
       this.closeDropdown();
-    } else if (!this.state.isOpen && keyCode === KeyCode.KEY_ESCAPE && this.hasEscPressed) {
-      this.hasEscPressed = false;
-      if (this.props.resetComboboxValue) {
-        this.props.resetComboboxValue();
-      }
-      this.setState({
-        hasSearchChanged: false,
-        searchValue: '',
-      });
-      event.stopPropagation();
-    } else if (keyCode === KeyCode.KEY_ESCAPE && this.props.allowClear) {
-      this.hasEscPressed = false;
+    } else if (keyCode === KeyCode.KEY_ESCAPE && this.props.allowClear && hasValue) {
       if (this.props.resetComboboxValue) {
         this.props.resetComboboxValue();
       }