+++ packages/terra-form-select/CHANGELOG.md
@@ -3,8 +3,9 @@
 ## Unreleased
 
 ## 6.52.0 - (December 11, 2023)
+
 * Added
-  * Added a condition to manage empty input field when the combo box is in focus.
+  * Added a condition to manage empty input field when the combobox is in focus.
 
 * Added
   * Added condition to announce the placeholder across all browsers.