+++ packages/terra-form-select/CHANGELOG.md
@@ -5,6 +5,9 @@
 ## 6.52.0 - (December 11, 2023)
 
 * Added
+  * Added a condition to manage empty input field when the combobox is in focus.
+
+* Added
   * Added condition to announce the placeholder across all browsers.
   * Added visual focus dashed border for `terra-form-select` list options.